[BOOK-67] refactor: 인증 전략에 Helper 패턴 도입
- 외부 의존성과 복잡한 로직을 별도의 Helper 클래스로 분리하였습니다:
- KakaoApiHelper: Kakao API 호출 책임 분리
- AppleJwtHelper: Apple JWT 파싱 로직 분리
- AuthTokenHelper: 토큰 관련 로직 분리
