+++ apis/src/main/kotlin/org/yapp/apis/auth/usecase/AuthUseCase.kt
@@ -1,39 +1,35 @@
 package org.yapp.apis.auth.usecase
 
 import org.springframework.stereotype.Service
+import org.yapp.apis.auth.dto.AuthCredentials
 import org.yapp.apis.auth.dto.response.TokenPairResponse
 import org.yapp.apis.auth.dto.response.UserProfileResponse
-import org.yapp.apis.auth.dto.AuthCredentials
+import org.yapp.apis.auth.helper.AuthTokenHelper
 import org.yapp.apis.auth.service.SocialAuthService
-import org.yapp.apis.auth.service.TokenService
 import org.yapp.apis.auth.service.UserAuthService
-import org.yapp.gateway.jwt.JwtTokenService
 import java.util.*
 
 @Service
 class AuthUseCase(
     private val socialAuthService: SocialAuthService,
     private val userAuthService: UserAuthService,
-    private val tokenService: TokenService,
-    private val jwtTokenService: JwtTokenService
+    private val authTokenHelper: AuthTokenHelper
 ) {
 
     fun signIn(credentials: AuthCredentials): TokenPairResponse {
         val strategy = socialAuthService.resolve(credentials)
         val userInfo = strategy.authenticate(credentials)
-
         val user = userAuthService.findOrCreateUser(userInfo)
-        return generateTokenPair(user.id!!)
+        return authTokenHelper.generateTokenPair(user.id!!)
     }
 
     fun refreshToken(refreshToken: String): TokenPairResponse {
-        val userId = jwtTokenService.getUserIdFromToken(refreshToken)
-        tokenService.validateRefreshTokenOrThrow(userId, refreshToken)
-        return generateTokenPair(userId)
+        val userId = authTokenHelper.validateAndGetUserIdFromRefreshToken(refreshToken)
+        return authTokenHelper.generateTokenPair(userId)
     }
 
     fun signOut(userId: UUID) {
-        tokenService.delete(userId)
+        authTokenHelper.deleteToken(userId)
     }
 
     fun getUserProfile(userId: UUID): UserProfileResponse {
@@ -47,15 +43,6 @@ class AuthUseCase(
     }
 
     fun getUserIdFromAccessToken(accessToken: String): UUID {
-        return jwtTokenService.getUserIdFromToken(accessToken)
-    }
-
-    private fun generateTokenPair(userId: UUID): TokenPairResponse {
-        val accessToken = jwtTokenService.generateAccessToken(userId)
-        val refreshToken = jwtTokenService.generateRefreshToken(userId)
-        val expiration = jwtTokenService.getRefreshTokenExpiration()
-
-        tokenService.save(userId, refreshToken, expiration)
-        return TokenPairResponse.of(accessToken, refreshToken)
+        return authTokenHelper.getUserIdFromAccessToken(accessToken)
     }
 }
